Mejoro doc en swagger con descripciones, ejemplos, etc. Agrego migraciones v1 y v2

+++ src/main/java/com/kenti/antezana/sistema_de_gestion_reservas/dto/response/ReservaRes.java
@@ -2,6 +2,7 @@ package com.kenti.antezana.sistema_de_gestion_reservas.dto.response;
 
 import com.kenti.antezana.sistema_de_gestion_reservas.model.EstadoReserva;
 import com.kenti.antezana.sistema_de_gestion_reservas.model.TipoDeEntrada;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record ReservaRes(
@@ -11,7 +12,7 @@ public record ReservaRes(
     TipoDeEntrada tipoDeEntrada,
     Long clienteId,
     EstadoReserva estadoReserva,
-    double montoTotal,
+    BigDecimal montoTotal,
     boolean bonificado
 ) {
 }
